progress
Former-commit-id: 180d4bcbb8bdb1417cf1e208653d16da3d2570e4 [formerly 974dc49960c8489f1e859405280a49f9cd51dc4c]
Former-commit-id: 31d664b171a5078fc427d2e4b3137323386f1f4c

+++ src/org/pflager/redbook/dpp/tess.java
@@ -97,7 +97,8 @@ public class tess extends glut {
 	 */
 	void combineCallback(double coords[3],
 	                     double *vertex_data[4],
-	                     GLfloat weight[4], double **dataOut )
+	                     double weight[4],
+	                     double **dataOut )
 	{
 	   double *vertex;
 	   int i;
